event-consumer/requirements.txt
kafka-python
redis
python-dateutil

